Bug 3500 : 64 bit integers for Gibson solver, expose Expression evaluation message (better reporting)
svn path=/vcell/trunk/numerics/; revision=12944


Former-commit-id: 8f94eb551e9faf5bb7bdf21d467667ec4439a21b

+++ Stochastic/VCellStoch/StochModel.h
@@ -3,6 +3,7 @@
 
 #include <string>
 #include <vector>
+#include <stdint.h>
 #include "StochVar.h"
 #include "Jump.h"
 
@@ -43,7 +44,7 @@ class StochModel
 	   int SEED;//random seed
 	   vector <StochVar*> listOfVars;//list of stochasic variables in the model
 	   vector <string> listOfVarNames;//list of variable names 
-	   vector <int> listOfIniValues;//keep original values for multiple trials
+	   vector <uint64_t> listOfIniValues;//keep original values for multiple trials
 	   vector <Jump*> listOfProcesses;//keep original values for multiple trials
 	   vector <string> listOfProcessNames;//list of process names
 };
